looping all scene in one day

+++ app/ftpSidebaPost.py
@@ -2,8 +2,9 @@ from ftplib import FTP
 import os
 from datetime import datetime
 import shutil
-def downloadFile():
-
+def downloadFile(liScene):
+	boolScene = False
+	print "Sudah diproses " + str(liScene)
 	tupDate = datetime.now()
 	print tupDate.year
 	print tupDate.strftime('%j')
@@ -22,6 +23,11 @@ def downloadFile():
 		ftp.cwd(str(level))
 		for scene in ftp.nlst():
 			print scene
+			if scene in liScene:
+				print "scene " + str(scene) + " sudah diproses"
+				continue;
+
+			boolScene = True
 			if(os.path.exists('C:/data/banjir/postFlood/'+ scene)):
 				shutil.rmtree('C:/data/banjir/postFlood/'+ scene)
 			os.makedirs('C:/data/banjir/postFlood/'+ scene)
@@ -45,19 +51,19 @@ def downloadFile():
 			ftp.cwd("../")
 		ftp.cwd("../")
 
-	os.chdir('C:/data/banjir/postFlood/'+ scene)
+	if(boolScene == True):
+		os.chdir('C:/data/banjir/postFlood/'+ scene)
+		for filename in os.listdir('C:/data/banjir/postFlood/'+ scene):
+			productID = filename.split(".")[0]
+			extension = filename.split(".")[1]
+			unique = filename.split(".")[0].split("_")[7]
 
-	for filename in os.listdir('C:/data/banjir/postFlood/'+ scene):
-		productID = filename.split(".")[0]
-		extension = filename.split(".")[1]
-		unique = filename.split(".")[0].split("_")[7]
+			os.rename(filename, scene + "_" + unique + "." + extension)
+			# print productID
+			# print extension
+			# print unique
 
-		os.rename(filename, scene + "_" + unique + "." + extension)
-		# print productID
-		# print extension
-		# print unique
-
-	return scene
+	return scene, boolScene
 
 # if __name__ == '__main__':
